+++ app/Ajax/App/Meeting/Remitment/Financial.php
@@ -142,6 +142,12 @@ class Financial extends CallableClass
 
     public function addRemitment()
     {
+        if($this->session->closed)
+        {
+            $this->notify->warning(trans('meeting.warnings.session.closed'));
+            return $this->response;
+        }
+
         $members = $this->remitmentService->getSubscriptions($this->pool);
         $title = trans('tontine.loan.titles.add');
         $content = $this->view()->render('tontine.pages.meeting.remitment.add')
@@ -168,6 +174,12 @@ class Financial extends CallableClass
      */
     public function saveRemitment(array $formValues)
     {
+        if($this->session->closed)
+        {
+            $this->notify->warning(trans('meeting.warnings.session.closed'));
+            return $this->response;
+        }
+
         $values = $this->validator->validateItem($formValues);
 
         $this->remitmentService->saveFinancialRemitment($this->pool,
@@ -185,6 +197,12 @@ class Financial extends CallableClass
      */
     public function deleteRemitment(int $subscriptionId)
     {
+        if($this->session->closed)
+        {
+            $this->notify->warning(trans('meeting.warnings.session.closed'));
+            return $this->response;
+        }
+
         $this->remitmentService->deleteFinancialRemitment($this->pool, $this->session, $subscriptionId);
         // $this->notify->success(trans('session.remitment.deleted'), trans('common.titles.success'));
